// src/services/api.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Upload: { input: any; output: any; }
};

export enum CacheControlScope {
  Private = 'PRIVATE',
  Public = 'PUBLIC'
}

export type Character = {
  __typename: 'Character';
  /** Time at which the character was created in the database. */
  created: Maybe<Scalars['String']['output']>;
  /** Episodes in which this character appeared. */
  episode: Array<Maybe<Episode>>;
  /** The gender of the character ('Female', 'Male', 'Genderless' or 'unknown'). */
  gender: Maybe<Scalars['String']['output']>;
  /** The id of the character. */
  id: Maybe<Scalars['ID']['output']>;
  /**
   * Link to the character's image.
   * All images are 300x300px and most are medium shots or portraits since they are intended to be used as avatars.
   */
  image: Maybe<Scalars['String']['output']>;
  /** The character's last known location */
  location: Maybe<Location>;
  /** The name of the character. */
  name: Maybe<Scalars['String']['output']>;
  /** The character's origin location */
  origin: Maybe<Location>;
  /** The species of the character. */
  species: Maybe<Scalars['String']['output']>;
  /** The status of the character ('Alive', 'Dead' or 'unknown'). */
  status: Maybe<Scalars['String']['output']>;
  /** The type or subspecies of the character. */
  type: Maybe<Scalars['String']['output']>;
};

export type Characters = {
  __typename: 'Characters';
  info: Maybe<Info>;
  results: Maybe<Array<Maybe<Character>>>;
};

export type Episode = {
  __typename: 'Episode';
  /** The air date of the episode. */
  air_date: Maybe<Scalars['String']['output']>;
  /** List of characters who have been seen in the episode. */
  characters: Array<Maybe<Character>>;
  /** Time at which the episode was created in the database. */
  created: Maybe<Scalars['String']['output']>;
  /** The code of the episode. */
  episode: Maybe<Scalars['String']['output']>;
  /** The id of the episode. */
  id: Maybe<Scalars['ID']['output']>;
  /** The name of the episode. */
  name: Maybe<Scalars['String']['output']>;
};
// ...
export type Location = {
  __typename: 'Location';
  /** Time at which the location was created in the database. */
  created: Maybe<Scalars['String']['output']>;
  /** The dimension in which the location is located. */
  dimension: Maybe<Scalars['String']['output']>;
  /** The id of the location. */
  id: Maybe<Scalars['ID']['output']>;
  /** The name of the location. */
  name: Maybe<Scalars['String']['output']>;
  /** List of characters who have been last seen in the location. */
  residents: Array<Maybe<Character>>;
  /** The type of the location. */
  type: Maybe<Scalars['String']['output']>;
};
// ...
export type AddSceneViewCharacterFragment = { __typename: 'Character', id: string | null, name: string | null, image: string | null, status: string | null };

export type AddSceneViewLocationFragment = { __typename: 'Location', id: string | null, name: string | null, type: string | null };